requirements.txt

Flask==3.1.2
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
gunicorn==20.1.0
pysicopg2-binary




